TYG-1: screens mockup
