+++ .gitignore
@@ -35,4 +35,4 @@ yarn-error.*
 *.tsbuildinfo
 
 # package
-paackage-lock.json
+package-lock.json
