+++ package.json
@@ -12,6 +12,8 @@
     "@expo-google-fonts/oswald": "^0.2.3",
     "@expo-google-fonts/rubik": "^0.2.3",
     "@expo/metro-runtime": "~3.1.3",
+    "@react-navigation/bottom-tabs": "^6.5.20",
+    "@react-navigation/drawer": "^6.6.15",
     "@react-navigation/native": "^6.1.17",
     "@react-navigation/native-stack": "^6.9.26",
     "babel-plugin-module-resolver": "^5.0.2",
@@ -26,7 +28,8 @@
     "react-native-reanimated": "~3.6.2",
     "react-native-safe-area-context": "4.8.2",
     "react-native-screens": "~3.29.0",
-    "react-native-web": "~0.19.6"
+    "react-native-web": "~0.19.6",
+    "react-native-gesture-handler": "~2.14.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
