+++ screens/ProfileScreen.js
@@ -1,0 +1,11 @@
+import { View, Text } from 'react-native';
+
+const ProfileScreen = () => {
+    return (
+        <View>
+            <Text>ProfileScreen</Text>
+        </View>
+    );
+};
+
+export default ProfileScreen;
